ai_bias_corpus_toolkit/requirements.txt
pandas>=2.2.0
feedparser>=6.0.11
requests>=2.31.0
python-dateutil>=2.8.2
dateparser>=1.2.0
beautifulsoup4>=4.12.3
trafilatura>=1.7.0
tqdm>=4.66.4
PyYAML>=6.0.2
rapidfuzz>=3.9.0
lxml>=4.9.3
lxml_html_clean>=0.2.0
